Uc1 regression testing (#190)
* Regression testing fixes

* Regression Testing fixes

* Regression Testing

Reduce debug log statements
Correct units for estimation exception throws

* UC1-Regression-testing: Fix typo in log statement.

+++ streets_utils/streets_vehicle_list/src/message_processors/all_stop_status_intent_processor.cpp
@@ -129,7 +129,7 @@ namespace streets_vehicles {
 		}
 
 		if (payload.FindMember("depart_pos")->value.IsInt64()){
-			vehicle._departure_position = payload["depart_pos"].GetInt();
+			vehicle._departure_position = payload["depart_pos"].GetInt64();
 		} else{
 			throw status_intent_processing_exception("The \"depart_pos\" " + vehicle._id + " is missing/incorrect in received update!");
